+++ src/Hex.h
@@ -4,5 +4,5 @@
 #define self Hex
 
 sdef(String, ToString, u64 hex);
-overload sdef(int, ToInteger, char c);
-overload sdef(long, ToInteger, String s);
+overload sdef(s8, ToInteger, char c);
+overload sdef(s64, ToInteger, String s);
